+++ src/test/resources/org/sonar/plugins/delphi/symbol/namespaces/Unit1.pas
@@ -1,0 +1,15 @@
+unit Unit1;
+
+{This is a sample Delphi file.}
+
+interface
+  procedure UnqualifiedUnit1Proc(Argument: String);
+
+implementation
+
+procedure UnqualifiedUnit1Proc(Argument: String);
+begin
+  // Do nothing
+end;
+
+end.