+++ src/main/java/org/sonar/plugins/delphi/preprocessor/search/SearchPath.java
@@ -1,14 +1,11 @@
 package org.sonar.plugins.delphi.preprocessor.search;
 
 import java.nio.file.Path;
-import java.util.Set;
 import javax.annotation.Nullable;
 
 public interface SearchPath {
   @Nullable
   Path search(String filename, Path startPath);
-
-  Set<Path> sourceFiles();
 
   static SearchPath create(Iterable<Path> searchDirectories) {
     return new DefaultSearchPath(searchDirectories);