+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Recursion bug in import resolution.
+
 ## [0.16.1] - 2020-03-26
 
 ### Fixed