+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.17.0] - 2020-04-27
+
 ### Added
 
 - `UnusedImportsRule` analysis rule, which flags imports that aren't used in the file.
@@ -371,7 +373,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - False-positives in 13 rules.
 - Logic errors in existing "semantic analysis" that caused scan failures.
 
-[Unreleased]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.16.2...HEAD
+[Unreleased]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.17.0...HEAD
+[0.17.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.16.2...v0.17.0
 [0.16.2]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.16.1...v0.16.2
 [0.16.1]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.16.0...v0.16.1
 [0.16.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.15.0...v0.16.0